Split IGraphicsContext from IPlatformWindow.
