+++ OpenRA.Game/Graphics/SpriteRenderer.cs
@@ -43,10 +43,10 @@ namespace OpenRA.Graphics
 					sheets[i] = null;
 				}
 
-				renderer.Window.SetBlendMode(currentBlend);
+				renderer.Context.SetBlendMode(currentBlend);
 				shader.PrepareRender();
 				renderer.DrawBatch(vertices, nv, PrimitiveType.TriangleList);
-				renderer.Window.SetBlendMode(BlendMode.None);
+				renderer.Context.SetBlendMode(BlendMode.None);
 
 				nv = 0;
 				ns = 0;
@@ -131,10 +131,10 @@ namespace OpenRA.Graphics
 		public void DrawVertexBuffer(IVertexBuffer<Vertex> buffer, int start, int length, PrimitiveType type, Sheet sheet, BlendMode blendMode)
 		{
 			shader.SetTexture("Texture0", sheet.GetTexture());
-			renderer.Window.SetBlendMode(blendMode);
+			renderer.Context.SetBlendMode(blendMode);
 			shader.PrepareRender();
 			renderer.DrawBatch(buffer, start, length, type);
-			renderer.Window.SetBlendMode(BlendMode.None);
+			renderer.Context.SetBlendMode(BlendMode.None);
 		}
 
 		// For RGBAColorRenderer
